Tests files moved to tests folders
