+++ src/components/tests/dispenser-item.test.tsx
@@ -1,12 +1,10 @@
-import {
-  ListItem,
-} from '@material-ui/core';
+import { ListItem } from '@material-ui/core';
 import { createMount } from '@material-ui/core/test-utils';
 import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import React from 'react';
-import fakeDispenserStates from '../mock/data/dispenser-states';
-import DispenserItem from './dispenser-item';
-import DispenserPanel from './dispensers-panel';
+import fakeDispenserStates from '../../mock/data/dispenser-states';
+import DispenserItem from '../dispenser-item';
+import DispenserPanel from '../dispensers-panel';
 
 const mount = createMount();
 
@@ -17,7 +15,7 @@ beforeEach(() => {
 });
 
 it('lists dispensers queued requests ID', () => {
-  Object.keys(dispenserStates).map( (guid, index) => {
+  Object.keys(dispenserStates).map((guid, index) => {
     const state = dispenserStates[guid];
 
     const root = mount(<DispenserItem dispenserState={state} />);
