+++ src/components/tests/door-item.test.tsx
@@ -1,9 +1,9 @@
 import { createMount } from '@material-ui/core/test-utils';
 import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import React from 'react';
-import buildingMap from '../mock/data/building-map';
-import fakeDoorStates from '../mock/data/door-states';
-import DoorItem from './door-item';
+import buildingMap from '../../mock/data/building-map';
+import fakeDoorStates from '../../mock/data/door-states';
+import DoorItem from '../door-item';
 
 const mount = createMount();
