+++ src/components/tests/doors-panel.test.tsx
@@ -1,11 +1,11 @@
 import { createMount } from '@material-ui/core/test-utils';
 import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import React from 'react';
-import buildingMap from '../mock/data/building-map';
-import fakeDoorStates from '../mock/data/door-states';
-import FakeTransport from '../mock/fake-transport';
-import DoorItem from './door-item';
-import DoorsPanel from './doors-panel';
+import buildingMap from '../../mock/data/building-map';
+import fakeDoorStates from '../../mock/data/door-states';
+import FakeTransport from '../../mock/fake-transport';
+import DoorItem from '../door-item';
+import DoorsPanel from '../doors-panel';
 
 const mount = createMount();
