+++ src/components/tests/main-menu.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import MainMenu from './main-menu';
+import MainMenu from '../main-menu';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
